# services/api-reservas/openapi.yaml
openapi: 3.1.0
info:
  title: CoworkReserve - Reservation API
  version: 1.0.0
  description: >
    API REST transaccional para gestionar reservas. Autenticación mediante Bearer JWT (Keycloak). Tiempos en ISO 8601.
servers:
  - url: http://localhost:3001
security:
  - bearerAuth: []
...
paths:
...
  /rooms/{id}:
    get:
      tags: [Rooms]
      summary: Obtener sala por ID
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '200':
          description: Sala
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags: [Rooms]
      summary: Modificar una sala
      description: Actualiza los datos de una sala existente.
      parameters:
        - $ref: '#/components/parameters/RoomId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateInput'
      responses:
        '200':
          description: Sala actualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Room'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Conflicto — ya existe otra sala con ese nombre
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    delete:
      tags: [Rooms]
      summary: Eliminar una sala
      description: >
        Elimina una sala del sistema. Las reservas asociadas a esta sala
        también serán eliminadas (cascade delete).
      parameters:
        - $ref: '#/components/parameters/RoomId'
      responses:
        '204':
          description: Sala eliminada exitosamente
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    RoomId:
      in: path
      name: id
      required: true
      schema:
        type: string
      description: Identificador de la sala
  responses:
    Unauthorized:
      description: No autenticado o token inválido
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Recurso no encontrado
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    BadRequest:
      description: Solicitud inválida
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
...
    Room:
      type: object
      required:
        - id
        - name
        - capacity
        - location
        - createdAt
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        capacity:
          type: integer
          minimum: 1
        location:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
    RoomUpdateInput:
      type: object
      properties:
        name:
          type: string
          description: Nombre único de la sala
        description:
          type: string
          nullable: true
        capacity:
          type: integer
          minimum: 1
          description: Número máximo de personas
        location:
          type: string
          nullable: true
          description: Ubicación física o piso
      description: >
        Todos los campos son opcionales. Solo se actualizarán los campos
        proporcionados.
    Problem:
      type: object
      description: RFC 9457 JSON Problem Details
      properties:
        type:   { type: string, format: uri, nullable: true }
        title:  { type: string }
        status: { type: integer }
        detail: { type: string, nullable: true }
        instance: { type: string, format: uri, nullable: true }
      required: [title, status]